Introduced a debug mode where you can send commands as text from a hyper terminal (eg. Arduino Serial Monitor).
The commands are sent like: "d0 0 0 127 63", which gets the goal keeper to rotate. There should be no space after the d but one space between each of the arguments. A space at the end is optional.

This introduced a change in the code so that we don't have to send a c continuously to get the regulators to update. This means that the regulators will update more often, but I don't think that is a problem (in that case, add a delay/time check). This shouldn't cause any problems with the normal server connection but if it does, this commit is a good place to start looking at.